Extract spec matchers into separate files
